Create MUS 424
This code is the result of my MUS 424 course/Research from a while back. You can pretty much ignore the first 300 lines as it only corresponds to the IGL research I was doing at the time. Very basic stuff here. Each circle is assigned musical characteristics based on the coordinate of its center in the complex plane. Higher/lower, positive/negative real values are associated with higher/lower pitches and higher/lower complex values are associated with louder/softer dynamics. Each circle is played in the order in which it is generated by the tree structure algorithm. It would be possible to utilize multiple cores to play multiple branches of the tree it this case, which we could also do for weather data if we choose.